Moved SimpleDrink to drink package
