+++ src/main/java/de/othr/sw/benjamineder/barmanagement/application/drink/dao/SimpleDrinkRepository.java
@@ -1,6 +1,6 @@
-package de.othr.sw.benjamineder.barmanagement.application.simpledrink.dao;
+package de.othr.sw.benjamineder.barmanagement.application.drink.dao;
 
-import de.othr.sw.benjamineder.barmanagement.application.simpledrink.entity.SimpleDrink;
+import de.othr.sw.benjamineder.barmanagement.application.drink.entity.SimpleDrink;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
