+++ src/main/java/de/othr/sw/benjamineder/barmanagement/application/drink/entity/SimpleDrinkType.java
@@ -1,0 +1,6 @@
+package de.othr.sw.benjamineder.barmanagement.application.drink.entity;
+
+public enum SimpleDrinkType {
+
+  BEER, WINE, LIQUOR, NON_ALCOHOLIC
+}
