--- src/main/java/de/othr/sw/benjamineder/barmanagement/application/simpledrink/entity/SimpleDrinkType.java
@@ -1,6 +1,0 @@
-package de.othr.sw.benjamineder.barmanagement.application.simpledrink.entity;
-
-public enum SimpleDrinkType {
-
-  BEER, WINE, LIQUOR, NON_ALCOHOLIC
-}
